requirements.txt
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/nightly/cu128

hydra-core==1.3.2
omegaconf